requirements.txt
openai
faiss-cpu
fastapi
uvicorn
python-dotenv
tqdm
PyMuPDF
numpy
jinja2               # for HTML templating
aiofiles             # for static file serving
streamlit